Separated qdlple gui from main, updated YAML loader and file to keep readers/aux-ctrls as lists
